Fixed indentation and moved to seperate file
I like the function, and please do tell me if I'm wrong in calling it the taylor series of sine.
Please be more aware of your indentation next time however.
